requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
async-timeout==3.0.1
attrs==21.4.0
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.0.12
feedparser==6.0.10
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
py-cord==2.0.0rc1
python-dotenv==0.20.0
sgmllib3k==1.0.0
typing_extensions==4.2.0
yarl==1.7.2